+++ test/commands/is_version_prepared_test.dart
@@ -11,6 +11,7 @@ import 'package:gg_console_colors/gg_console_colors.dart';
 import 'package:gg_git/gg_git_test_helpers.dart';
 import 'package:gg_version/gg_version.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:path/path.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:test/test.dart';
 
@@ -70,7 +71,6 @@ void main() async {
             },
           );
         });
-
         group('and log the required versions', () {
           test('when versions are not the next increment', () async {
             // Assume the published version is 2.0.0
@@ -93,7 +93,7 @@ void main() async {
             expect(result, isFalse);
             expect(
               messages.last,
-              darkGray('$versions must one of the following:'
+              darkGray('$versions must be one of the following:'
                   '\n- 2.0.1'
                   '\n- 2.1.0'
                   '\n- 3.0.0'),
@@ -103,7 +103,7 @@ void main() async {
       });
 
       group('should return true', () {
-        group('when README.md and pubspec.yaml have the same version', () {
+        group('when CHANGELOg.md and pubspec.yaml have the same version', () {
           test('and the version is the next increment', () async {
             // Assume the published version is 2.0.0
             when(() => publishedVersion.get(ggLog: ggLog, directory: d))
@@ -124,6 +124,40 @@ void main() async {
               expect(result, isTrue);
               expect(messages.isEmpty, isTrue);
             }
+          });
+        });
+
+        group('when CHANGELOg.md and pubspec.yaml have not the same version',
+            () {
+          group('but CHANGELOG.md has an ## "Unreleased" headline', () {
+            test('and treatUnpublishedAsOk is true', () async {
+              // Assume the published version is 2.0.0
+              when(() => publishedVersion.get(ggLog: ggLog, directory: d))
+                  .thenAnswer((_) async => Version(2, 0, 0));
+
+              // Assume the locally configured version is 3.0.0
+              await addAndCommitVersions(
+                d,
+                pubspec: '2.1.0',
+                changeLog: '2.0.0',
+                gitHead: '2.0.0',
+              );
+
+              // Prepare CHANGELOG.md
+              File(join(d.path, 'CHANGELOG.md')).writeAsStringSync(
+                '# Changelog\n\n'
+                '## Unreleased\n\n- Message 1\n\n'
+                '## 3.0.0\n\n- Message 2\n',
+              );
+
+              final result = await isVersionPrepared.get(
+                ggLog: ggLog,
+                directory: d,
+                treatUnpublishedAsOk: true,
+              );
+              expect(result, isTrue);
+              expect(messages.isEmpty, isTrue);
+            });
           });
         });
       });
@@ -159,11 +193,20 @@ void main() async {
 
       group('should print »❌ Version is prepared«', () {
         group('and throw an error description', () {
-          test('when the versions do not match', () async {
+          test('when the version in pubspec is not an increment', () async {
+            // Assume the published version is 2.0.0
+            when(
+              () => publishedVersion.get(
+                ggLog: any(named: 'ggLog'),
+                directory: any(named: 'directory'),
+              ),
+            ).thenAnswer((_) async => Version(2, 0, 0));
+
+            // The local version is 2.5.0
             await addAndCommitVersions(
               d,
-              pubspec: '2.0.0',
-              changeLog: '3.0.0',
+              pubspec: '2.5.0', // Not an increment
+              changeLog: '2.5.0', // Not an increment
               gitHead: '2.0.1',
             );
 
@@ -182,10 +225,10 @@ void main() async {
             expect(messages[0], contains('⌛️ Version is prepared'));
             expect(messages[1], contains('❌ Version is prepared'));
 
-            expect(
-              exceptionMessage,
-              contains(darkGray('$versions must be the same.')),
-            );
+            expect(exceptionMessage, contains('must be one of the following'));
+            expect(exceptionMessage, contains('2.0.1'));
+            expect(exceptionMessage, contains('2.1.0'));
+            expect(exceptionMessage, contains('3.0.0'));
           });
         });
       });